requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
discord.py==2.3.2
filelock==3.15.3
frozenlist==1.4.1
idna==3.7
lxml==5.2.2
multidict==6.0.5
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
snscrape==0.7.0.20230622
soupsieve==2.5
urllib3==2.2.2
yarl==1.9.4
